MPU-9250 tests
MPU-9250 tests: using interrupts, read raw values from accelerometer,
magnetometer and gyroscope, and write them along with a time stamp to
the FRAM. Attempt to do this at 200 Hz.

There appears to be a weird issue where the loop only runs for a small
(and unpredictable) amount of time. Could be because of things not
happening fast enough? Need to try testing MPU-9250 at a lower update
frequency to find out.